refactor: refactoring bifunction dowload path in report abstract

+++ src/main/java/dev/rayan/report/ReportAbstractFile.java
@@ -5,15 +5,16 @@ import dev.rayan.enums.TransactionReportPeriod;
 
 import java.io.File;
 import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.Paths;
 import java.util.function.BiFunction;
 
 public abstract class ReportAbstractFile {
 
     private static final String USER_HOME = System.getProperty("user.home");
 
-    private static final BiFunction<String, String, File> DOWNLOAD_PATH_FUNCTION = (fileName, extension) -> new File(
-            String.format("%s/Downloads/%s%s", USER_HOME, fileName, extension)
-    );
+    private static final BiFunction<String, String, File> DOWNLOAD_PATH_FUNCTION = (fileName, extension) -> Paths.get(USER_HOME, "Downloads", fileName, extension).toFile();
 
     public abstract void createReport(TransactionReportResponse response, TransactionReportPeriod period) throws IllegalAccessException, IOException;
 
@@ -23,14 +24,26 @@ public abstract class ReportAbstractFile {
 
     public File createDownloadPath() {
 
+
         File downloadPath = DOWNLOAD_PATH_FUNCTION.apply(getFileName(), getExtension());
         int version = 1;
 
-        while (downloadPath.exists()) {
-            downloadPath.renameTo(DOWNLOAD_PATH_FUNCTION.apply(getFileName() + " (" + version + ")", getExtension()));
-            version++;
+        //arquivo existe
+        if (downloadPath.exists()) {
+
+            //está aberto
+            boolean isOpen = downloadPath.renameTo(downloadPath);
+
+
+            //está fechado
+            while (downloadPath.exists()) {
+                downloadPath.renameTo(DOWNLOAD_PATH_FUNCTION.apply(getFileName() + " (" + version + ")", getExtension()));
+                version++;
+            }
+
         }
 
+        //arquivo não existe ou já alterou
         return downloadPath;
     }
